requirements.txt
boto3==1.34.114
botocore==1.34.114
contourpy==1.2.1
cycler==0.12.1
fonttools==4.52.4
iniconfig==2.0.0
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
pyparsing==3.1.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
